separate file: InferNonNullAggConstraintTests
